Rename 'quarry.util' to 'quarry.utils'
