+++ quarry/utils/crypto.py
