+++ quarry/utils/types.py
